requirements.txt
altair==4.2.0
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.4
Pillow==9.3.0
scikit_learn==1.2.0
scipy==1.9.1
seaborn==0.11.2
streamlit==1.13.0
